Finished basics of configuration. Started inventory.
