+++ css/reset.css
@@ -3,7 +3,7 @@
     margin: 0;
     padding: 0;
     box-sizing: border-box;
-    outline: 1px solid #d84315;
+    /* outline: 1px solid #d84315; */
 }
 
 html {
