+++ views/inventory.html
@@ -1,6 +1,16 @@
-            <div class="header">
-                <h2>All Plants</h2>
-                <p>Your complete plant collection</p>
-            </div>
-            <div class="plants-grid" id="plants-grid"></div>
-    
+<div class="header">
+    <h2>Inventory</h2>
+    <p>Manage all your plants</p>
+</div>
+
+<div id="plant-inventory-table" class="table">
+                <div class="table-header">
+                    <div>Name</div>
+                    <div>Location</div>
+                    <div>Status</div>
+                    <div>Last Activity</div>
+                    <div></div>
+                </div>
+                
+              
+</div>
